Simplify redundant (ulong)(int) casts of Time values to (ulong)
Time.cs already defines an explicit operator ulong that goes straight
from Time's underlying Seconds, so routing through int first was dead
weight -- not scope-related to #51, just noticed in passing.

+++ ServiceSiteScheduling/Solutions/PlanGraph.cs
@@ -1230,7 +1230,7 @@ namespace ServiceSiteScheduling.Solutions
                         infrastructure.Sum(infra => (long)GetFlatDuration(infra));
                     if (leadingReversalCredit)
                     {
-                        flatDuration += (ulong)(int)Settings.TrackCrossingTime;
+                        flatDuration += (ulong)Settings.TrackCrossingTime;
                         leadingReversalCredit = false;
                     }
